Move QueryColumn, LoadingState to public.
- Move hasAllPermissions to User.ts to get rid of circular import
