+++ packages/components/src/internal/components/assay/RunPropertiesPanel.spec.tsx
@@ -10,7 +10,7 @@ import { CheckboxInput } from '../forms/input/CheckboxInput';
 import { SelectInput } from '../forms/input/SelectInput';
 
 import { QueryFormInputs } from '../forms/QueryFormInputs';
-import { QueryColumn } from '../base/models/model';
+import { QueryColumn } from '../../..';
 
 import { AssayWizardModel } from './models';
 import { RunPropertiesPanel } from './RunPropertiesPanel';
