+++ packages/components/src/internal/components/base/models/User.ts
@@ -1,8 +1,6 @@
 import { List, Record } from 'immutable';
 import { ActionURL, PermissionTypes, UserWithPermissions } from '@labkey/api';
 
-import { hasAllPermissions } from '../../../..';
-
 interface IUserProps extends Partial<UserWithPermissions> {
     permissionsList: List<string>;
 }
@@ -84,4 +82,30 @@ export class User extends Record(defaultUser) implements IUserProps {
     isAppAdmin(): boolean {
         return hasAllPermissions(this, [PermissionTypes.ApplicationAdmin], false);
     }
+}
+
+/**
+ * Determines if a user has all of the permissions given.  If the user has only some
+ * of these permissions, returns false.
+ * @param user the user in question
+ * @param perms the list of permission strings (See models/constants)
+ * @param checkIsAdmin boolean indicating if user.isAdmin should be used as a fallback check
+ */
+export function hasAllPermissions(user: User, perms: string[], checkIsAdmin = true): boolean {
+    let allow = false;
+
+    if (perms) {
+        const allPerms = user.get('permissionsList');
+
+        let hasAll = true;
+        for (let i = 0; i < perms.length; i++) {
+            if (allPerms.indexOf(perms[i]) === -1) {
+                hasAll = false;
+                break;
+            }
+        }
+        allow = hasAll || (checkIsAdmin && user.isAdmin);
+    }
+
+    return allow;
 }
