+++ packages/components/src/internal/components/domainproperties/actions.spec.ts
@@ -15,7 +15,7 @@
  */
 import { List } from 'immutable';
 
-import { QueryColumn } from '../base/models/model';
+import { QueryColumn } from '../../..';
 
 import {
     createFormInputId,
