+++ packages/components/src/internal/components/editable/EditableGrid.tsx
@@ -39,7 +39,7 @@ import { QueryInfoForm, QueryInfoFormProps } from '../forms/QueryInfoForm';
 import { MAX_EDITABLE_GRID_ROWS } from '../../constants';
 import { Grid, GridColumn } from '../base/Grid';
 import { GRID_CHECKBOX_OPTIONS, GRID_EDIT_INDEX, GRID_SELECTION_INDEX } from '../base/models/constants';
-import { QueryColumn, QueryGridModel } from '../base/models/model';
+import { QueryGridModel } from '../base/models/model';
 import { blurActiveElement, capitalizeFirstChar, caseInsensitive } from '../../util/utils';
 import { DeleteIcon } from '../base/DeleteIcon';
 import { Alert } from '../base/Alert';
@@ -49,6 +49,7 @@ import { BulkAddUpdateForm } from '../forms/BulkAddUpdateForm';
 
 import { AddRowsControl, AddRowsControlProps, PlacementType } from './Controls';
 import { Cell } from './Cell';
+import { QueryColumn } from '../../..';
 
 const COUNT_COL = new GridColumn({
     index: GRID_EDIT_INDEX,
