+++ packages/components/src/internal/components/forms/PageDetailHeader.tsx
@@ -16,9 +16,7 @@
 import React, { PureComponent, ReactNode } from 'react';
 import { PermissionTypes } from '@labkey/api';
 
-import { hasAllPermissions } from '../../util/utils';
-import { User } from '../../..';
-import { SVGIcon } from '../base/SVGIcon';
+import { hasAllPermissions, SVGIcon, User } from '../../..';
 
 import { FieldEditorOverlay, FieldEditorOverlayProps } from './FieldEditorOverlay';
