+++ packages/components/src/public/LoadingState.ts
@@ -1,0 +1,8 @@
+export enum LoadingState {
+    // The model has been initialized but not loaded
+    INITIALIZED = 'INITIALIZED',
+    // The model is currently loading
+    LOADING = 'LOADING',
+    // The model is loaded
+    LOADED = 'LOADED',
+}
